Extracted Packets, Plugin & Networking code into sub modules / external project
