+++ packets/pom.xml
@@ -1,0 +1,17 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+    <packaging>jar</packaging>
+    <parent>
+        <groupId>com.endersuite</groupId>
+        <artifactId>endersync</artifactId>
+        <version>5.0.0</version>
+    </parent>
+
+    <groupId>com.endersuite.endersync</groupId>
+    <artifactId>packets</artifactId>
+
+
+</project>
