+++ plugin/pom.xml
@@ -1,0 +1,59 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project xmlns="http://maven.apache.org/POM/4.0.0"
+         xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance"
+         xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
+    <modelVersion>4.0.0</modelVersion>
+    <packaging>jar</packaging>
+    <parent>
+        <groupId>com.endersuite</groupId>
+        <artifactId>endersync</artifactId>
+        <version>5.0.0</version>
+    </parent>
+
+    <groupId>com.endersuite.endersuite</groupId>
+    <artifactId>plugin</artifactId>
+
+    <dependencies>
+
+        <!-- JGroup -->
+        <dependency>
+            <groupId>org.jgroups</groupId>
+            <artifactId>jgroups</artifactId>
+            <version>5.1.6.Final</version>
+            <scope>provided</scope>
+        </dependency>
+
+        <!-- Endersync-packets -->
+        <dependency>
+            <groupId>com.endersuite.endersync</groupId>
+            <artifactId>packets</artifactId>
+            <version>5.0.0</version>
+            <scope>provided</scope>
+        </dependency>
+
+        <!-- Caffeine -->
+        <dependency>
+            <groupId>com.github.ben-manes.caffeine</groupId>
+            <artifactId>caffeine</artifactId>
+            <version>3.0.2</version>
+            <scope>provided</scope>
+        </dependency>
+
+        <!-- LibCore -->
+        <dependency>
+            <groupId>com.endersuite</groupId>
+            <artifactId>libcore</artifactId>
+            <version>1.4.0</version>
+            <scope>compile</scope>
+        </dependency>
+
+        <!-- JEventLoop -->
+        <dependency>
+            <groupId>de.maximilian-heidenreich</groupId>
+            <artifactId>jeventloop</artifactId>
+            <version>2.5.0</version>
+            <scope>provided</scope>
+        </dependency>
+
+    </dependencies>
+</project>
