+++ plugin/src/main/java/com/endersuite/endersync/exceptions/DuplicateModuleNameException.java
