+++ plugin/src/main/java/com/endersuite/endersync/exceptions/EnderSyncException.java
