+++ plugin/src/main/resources/jgroup_conf.xml
