+++ pom.xml
@@ -6,11 +6,18 @@
 
     <groupId>com.endersuite</groupId>
     <artifactId>endersync</artifactId>
+    <packaging>pom</packaging>
     <version>5.0.0</version>
+
+    <modules>
+        <module>plugin</module>
+        <module>packets</module>
+    </modules>
 
     <properties>
         <maven.compiler.source>8</maven.compiler.source>
         <maven.compiler.target>8</maven.compiler.target>
+        <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
     </properties>
 
     <dependencies>
@@ -31,27 +38,12 @@
             <scope>provided</scope>
         </dependency>
 
-        <!-- JGroup -->
-        <dependency>
-            <groupId>org.jgroups</groupId>
-            <artifactId>jgroups</artifactId>
-            <version>5.1.6.Final</version>
-            <scope>provided</scope>
-        </dependency>
-
-        <!-- Caffeine -->
-        <dependency>
-            <groupId>com.github.ben-manes.caffeine</groupId>
-            <artifactId>caffeine</artifactId>
-            <version>3.0.2</version>
-            <scope>provided</scope>
-        </dependency>
-
-        <!-- LibCore -->
+        <!-- Packify -->
         <dependency>
             <groupId>com.endersuite</groupId>
-            <artifactId>libcore</artifactId>
-            <version>1.4.0</version>
+            <artifactId>packify</artifactId>
+            <version>1.0.0</version>
+            <scope>provided</scope>
         </dependency>
 
         <!-- Database -->
@@ -61,14 +53,7 @@
             <version>1.0.0</version>
         </dependency>
 
-        <!-- JEventLoop -->
-        <dependency>
-            <groupId>de.maximilian-heidenreich</groupId>
-            <artifactId>jeventloop</artifactId>
-            <version>2.4.0</version>
-        </dependency>
-
-    </dependencies>
+   </dependencies>
 
     <build>
         <plugins>
@@ -83,7 +68,7 @@
                             <goal>shade</goal>
                         </goals>
                         <configuration>
-                            <minimizeJar>false</minimizeJar>
+                            <minimizeJar>true</minimizeJar>
                         </configuration>
                     </execution>
                 </executions>
