--- src/main/java/com/endersuite/endersync/events/core/PacketReceivedEvent.java
@@ -1,54 +1,0 @@
-package com.endersuite.endersync.events.core;
-
-import com.endersuite.endersync.networking.packets.APacket;
-import de.maximilianheidenreich.jeventloop.events.AbstractEvent;
-import lombok.Getter;
-import org.jgroups.Address;
-
-/**
- * Gets dispatched whenever an {@link APacket} was received over the network.
- *
- * @author Maximilian Vincent Heidenreich
- * @since 09.05.21
- */
-public class PacketReceivedEvent extends AbstractEvent<Void> {
-
-    // ======================   VARS
-
-    /**
-     * The received packet.
-     */
-    @Getter
-    private final APacket packet;
-
-    /**
-     * The sender of the packet.
-     */
-    @Getter
-    private final Address sender;
-
-    /**
-     * The recipient of the packet.
-     */
-    @Getter
-    private final Address recipient;
-
-
-    // ======================   CONSTRUCTOR
-
-    public PacketReceivedEvent(APacket packet) {
-        super((byte) 1);
-        this.packet = packet;
-        this.sender = packet.getSender();
-        this.recipient = packet.getRecipient();
-    }
-
-
-    // ======================   HELPERS
-
-    @Override
-    public String toString() {
-        return String.format("%s(%s)-s(%s)-r(%s)", this.getClass().getSimpleName(), getId().toString().split("-")[0], sender, recipient);
-    }
-
-}
