--- src/main/java/com/endersuite/endersync/networking/packets/APacket.java
@@ -1,79 +1,0 @@
-package com.endersuite.endersync.networking.packets;
-
-import lombok.Getter;
-import lombok.Setter;
-import org.jgroups.Address;
-
-import java.io.Serializable;
-import java.util.UUID;
-
-/**
- * A abstract packet implementation that can be sent over the network.
- *
- * @author Maximilian Vincent Heidenreich
- * @since 09.05.21
- */
-@Getter
-public abstract class APacket implements Serializable {
-
-    // ======================   VARS
-
-    /**
-     * A unique id identifying the packet.
-     */
-    @Setter
-    private final UUID id;
-
-    /**
-     * Timestamp at which the packet instance was created.
-     */
-    @Getter
-    private final long createdAt;
-
-    /**
-     * The sender of the packet (Only populated when received).
-     */
-    @Getter
-    @Setter
-    private Address sender;
-
-    /**
-     * The recipient of the packet (Only populated when received).
-     */
-    @Getter
-    @Setter
-    private Address recipient;
-
-    // ======================   CONSTRUCTOR
-
-    /**
-     * Create a new AbstractPacket with a random id.
-     */
-    public APacket() {
-        this(UUID.randomUUID());
-    }
-
-    /**
-     * Creates a new AbstractPacket with a specified id.
-     *
-     * @param id
-     *          The id to use.
-     */
-    public APacket(UUID id) {
-        this.id = id;
-        this.createdAt = System.currentTimeMillis();
-    }
-
-
-    // ======================   HELPERS
-
-    @Override
-    public String toString() {
-        if (sender != null && recipient != null)
-            return String.format("%s(%s)_s(%s)-r(%s)", this.getClass().getSimpleName(), getId().toString().split("-")[0], sender, recipient);
-        else
-            return String.format("%s(%s)", this.getClass().getSimpleName(), getId().toString().split("-")[0]);
-    }
-
-
-}
